https://www.hackerrank.com/challenges/bfsshortreach/submissions/code/41214706

+++ RoadsAndLibraries.cpp
@@ -29,7 +29,7 @@ int main() {
     int b;
     int cm;
     int cd;
-    int sum;
+    long long sum;
     
     
     cin >> q ;
